make setGamePath not fail if User/ doesn't exist

+++ app/domain/DolphinManager.js
@@ -55,8 +55,9 @@ export default class DolphinManager {
   }
 
   async resetDolphin() {
+    const platform = process.platform;
     const appPath = app.getAppPath();
-    const originalDolphinPath = path.join(appPath, "../app.asar.unpacked/app/dolphin");
+    const originalDolphinPath = process.env.APPIMAGE ? path.join(process.env.APPDIR, "resources/app.asar.unpacked/app/dolphin") : path.join(appPath, "../app.asar.unpacked/app/dolphin");
     log.info("Resetting dolphin");
     const userDataPath = app.getPath("userData");
     const targetPath = path.join(userDataPath, 'dolphin');
@@ -64,9 +65,14 @@ export default class DolphinManager {
 
     let isCopySuccess = false;
     try {
-      fs.removeSync(targetPath);
-      fs.copySync(originalDolphinPath, targetPath);
-      log.info("Dolphin was reset");
+      if (platform === "win32" || platform === "darwin" || process.env.APPIMAGE){
+        fs.removeSync(targetPath);
+        fs.copySync(originalDolphinPath, targetPath);
+      }
+      if (process.platform === "linux") {
+        const linuxUserDir = path.join(os.homedir(),".config", "SlippiPlayback");
+        fs.removeSync(linuxUserDir); // clear the User dir on linux
+      }
       isCopySuccess = true;
     } catch (err) {
       log.error("Failed to reset Dolphin, will try again with elevated permissions");
@@ -78,10 +84,9 @@ export default class DolphinManager {
         // TODO: ipc to trigger it in the main process? But I'm too lazy right now
         await sudoRemovePath(targetPath);
         fs.copySync(originalDolphinPath, targetPath);
-        log.info("Dolphin was reset");
         isCopySuccess = true;
       } catch (err) {
-        log.error("Failed to reset Dolphin, will try again with elevated permissions");
+        log.error("Failed to reset Dolphin with elevated permissions");
         log.error(err);
         throw new Error("Failed to reset Dolphin. You may need to reinstall the desktop app.");
       }
@@ -110,12 +115,21 @@ export default class DolphinManager {
     }
     try {
       const iniPath = path.join(dolphinPath, "Config", "Dolphin.ini");
-      const dolphinINI = ini.parse(fs.readFileSync(iniPath, 'utf-8'));
-      dolphinINI.General.ISOPath0 = fileDir;
-      const numPaths = dolphinINI.General.ISOPaths;
-      dolphinINI.General.ISOPaths = numPaths !== "0" ? numPaths : "1";
-      const newINI = ini.encode(dolphinINI);
-      fs.writeFileSync(iniPath, newINI);
+      if (fs.existsSync(iniPath)){
+        const dolphinINI = ini.parse(fs.readFileSync(iniPath, 'utf-8'));
+        dolphinINI.General.ISOPath0 = fileDir;
+        const numPaths = dolphinINI.General.ISOPaths;
+        dolphinINI.General.ISOPaths = numPaths !== "0" ? numPaths : "1";
+        const newINI = ini.encode(dolphinINI);
+        fs.writeFileSync(iniPath, newINI);
+      } else {
+        log.info("There isn't a User dir to write to.");
+        const configPath = path.join(dolphinPath, "Config");
+        const newINI = ini.encode({"General": {"ISOPath0": fileDir, "ISOPaths": 1}});
+        log.info("attempting to mkdir -p");
+        fs.mkdirpSync(configPath);
+        fs.writeFileSync(iniPath, newINI);
+      }
     } catch (err) {
       log.warn(`Failed to update the dolphin paths\n${err}`);
       throw err;
